+++ cheat-base/framework/pch.h
@@ -17,6 +17,8 @@
 #include <string>
 #include <vector>
 #include <algorithm>
+#include <filesystem>
+#include <fstream>
 #include <map>
 #include <unordered_set>
 #include <queue>
