+++ cheat-base/src/cheat-base/inject/load-library.h
@@ -1,0 +1,3 @@
+#pragma once
+
+bool LoadLibraryDLL(HANDLE hProc, const std::string& dllpath);
